+++ lib/vagrant-parallels/action/import.rb
@@ -24,7 +24,8 @@ module VagrantPlugins
           end
 
           # Linked clones are supported only for PD 11 and higher
-          if env[:machine].provider_config.linked_clone
+          # Linked clones are not supported in macvms
+          if env[:machine].provider_config.linked_clone and !Util::Common::is_macvm(env[:machine])
             # Linked clone creation should not be concurrent [GH-206]
             options[:snapshot_id] = env[:clone_snapshot_id]
             options[:linked] = true