+++ lib/vagrant-parallels/driver/base.rb
@@ -403,11 +403,13 @@ module VagrantPlugins
           end
         end
 
-        # Returns an IP of the virtual machine. It requires that Shared network
-        # adapter is configured for this VM and it obtains an IP via DHCP.
+        # Returns an IP of the virtual machine fetched from the DHCP lease file.
+        # It requires that Shared network adapter is configured for this VM
+        # and it obtains an IP via DHCP.
+        # Returns an empty string if the IP coudn't be determined this way.
         #
         # @return [String] IP address leased by DHCP server in "Shared" network
-        def read_guest_ip
+        def read_guest_ip_dhcp
           mac_addr = read_mac_address.downcase
           leases_file = '/Library/Preferences/Parallels/parallels_dhcp_leases'
           leases = {}
@@ -428,6 +430,16 @@ module VagrantPlugins
 
           # Get the most resent lease and return an associated IP
           leases.max_by { |_ip, lease_time| lease_time }.first
+        end
+
+        # Returns an IP of the virtual machine fetched from prlctl.
+        # Returns an empty string if the IP coudn't be determined this way.
+        #
+        # @return [String] IP address returned by `prlctl list -f` command
+        def read_guest_ip_prlctl
+          vm_info = json { execute_prlctl('list', @uuid, '--full', '--json') }
+          ip = vm_info.first.fetch('ip_configured', '')
+          ip == '-' ? '' : ip
         end
 
         # Returns path to the Parallels Tools ISO file.
@@ -757,11 +769,26 @@ module VagrantPlugins
           end
         end
 
-        # Reads the SSH IP of this VM.
+        # Reads the SSH IP of this VM from DHCP lease file or from `prlctl list`
+        # command - whatever returns a non-empty result first.
+        # The method with DHCP does not work for *.macvm VMs on Apple M-series Macs,
+        # so we try both sources here.
         #
         # @return [String] IP address to use for SSH connection to the VM.
         def ssh_ip
-          read_guest_ip
+          5.times do
+            ip = read_guest_ip_dhcp
+            return ip unless ip.empty?
+
+            ip = read_guest_ip_prlctl
+            return ip unless ip.empty?
+
+            sleep 2
+          end
+
+          # We didn't manage to determine IP - return nil and
+          # expect SSH client to do a retry
+          return nil
         end
 
         # Reads the SSH port of this VM.