+++ lib/vagrant-parallels/plugin.rb
@@ -103,6 +103,16 @@ module VagrantPlugins
         SyncedFolder
       end
 
+      synced_folder(:parallels_macvm) do
+        require_relative 'synced_folder_macvm'
+        SyncedFolderMacVM
+      end
+
+      synced_folder_capability(:parallels_macvm, "mount_name") do
+        require_relative "cap/mount_options"
+        SyncedFolderCap::MountOptions
+      end
+
       synced_folder_capability(:parallels, "mount_name") do
         require_relative "cap/mount_options"
         SyncedFolderCap::MountOptions
@@ -168,6 +178,7 @@ module VagrantPlugins
 
     module Util
       autoload :CompileForwardedPorts, File.expand_path('../util/compile_forwarded_ports', __FILE__)
+      autoload :Common, File.expand_path('../util/common', __FILE__)
     end
   end
 end