+++ test/unit/synced_folder_test.rb
@@ -9,6 +9,7 @@ describe VagrantPlugins::Parallels::SyncedFolder do
     double('machine').tap do |m|
       allow(m).to receive_messages(provider_config: VagrantPlugins::Parallels::Config.new)
       allow(m).to receive_messages(provider_name: :parallels)
+      allow(VagrantPlugins::Parallels::Util::Common).to receive(:is_macvm).with(m).and_return(false)
     end
   end
 
@@ -26,6 +27,11 @@ describe VagrantPlugins::Parallels::SyncedFolder do
 
     it 'should not be with another provider' do
       allow(machine).to receive_messages(provider_name: :virtualbox)
+      expect(subject).not_to be_usable(machine)
+    end
+
+    it 'should not be with macvm guest' do
+      allow(VagrantPlugins::Parallels::Util::Common).to receive(:is_macvm).with(machine).and_return(true)
       expect(subject).not_to be_usable(machine)
     end